refac: equation code
