+++ backend/RacerHub.cs
@@ -1,5 +1,7 @@
 using System.Text.Json;
+using equation;
 using Microsoft.AspNetCore.SignalR;
+using models;
 
 namespace hub;
 
@@ -64,12 +66,14 @@ public class RacerHub : Hub
 
         Player p = lobby[id];
         lobby.Remove(id);
+
+        await Groups.RemoveFromGroupAsync(Context.ConnectionId, gameId);
+
         if (lobby.Count == 0)
         {
             lobbies.Remove(gameId);
+            return;
         }
-
-        await Groups.RemoveFromGroupAsync(Context.ConnectionId, gameId);
 
         // TODO: change host if host leaves
         // if (p.isHost) {}
@@ -77,150 +81,17 @@ public class RacerHub : Hub
         syncPlayers(gameId);
     }
 
-    private int GetRandomInt(int min, int max)
-    {
-        Random random = new Random();
-        return random.Next(min, max + 1);
-    }
-
-    private int GenerateId()
-    {
-        return 0; //TODO
-    }
-
-    private Equation GenerateAddition()
-    {
-        int num1 = GetRandomInt(1, 20);
-        int num2 = GetRandomInt(1, 20);
-        return new Equation(GenerateId(), $"{num1} + {num2} = ?", num1 + num2);
-    }
-
-    private Equation GenerateSubtraction()
-    {
-        int answer = GetRandomInt(1, 20);
-        int num2 = GetRandomInt(1, 10);
-        int num1 = answer + num2;
-        return new Equation(GenerateId(), $"{num1} - {num2} = ?", answer);
-    }
-
-    private Equation GenerateMultiplication()
-    {
-        int num1 = GetRandomInt(1, 12);
-        int num2 = GetRandomInt(1, 12);
-        return new Equation(GenerateId(), $"{num1} × {num2} = ?", num1 * num2);
-    }
-
-    private Equation GenerateDivision()
-    {
-        int answer = GetRandomInt(1, 10);
-        int num2 = GetRandomInt(1, 10);
-        int num1 = answer * num2;
-        return new Equation(GenerateId(), $"{num1} ÷ {num2} = ?", answer);
-    }
-
-    private Equation GenerateEquation()
-    {
-        int operationType = GetRandomInt(1, 5);
-        switch (operationType)
-        {
-            case 1:
-                return GenerateAddition();
-            case 2:
-                return GenerateSubtraction();
-            case 3:
-                return GenerateMultiplication();
-            case 4:
-                return GenerateDivision();
-            default:
-                return GenerateAddition();
-        }
-    }
-
-    public Equation[] GenerateAllEquations(int count)
-    {
-        Equation[] equations = new Equation[count];
-        for (int i = 0; i < count; i++)
-        {
-            equations[i] = GenerateEquation();
-        }
-        return equations;
-    }
-
-    public void GenerateEquations(string gameId, string mode, int count)
-    {
-        Equation[] equations = GenerateAllEquations(count);
-        Game game = new Game(gameId, equations, new GameMode(mode, count));
-        games.Add(gameId, game);
-    }
-
     public void StartGame(string gameId, string mode)
     {
         GameMode selectedMode = JsonSerializer.Deserialize<GameMode>(mode)!;
         GenerateEquations(gameId, selectedMode.type, selectedMode.count);
     }
-}
 
-public class Player
-{
-    public int id { get; set; }
-    public int progress { get; set; }
-    public int score { get; set; }
-    public bool isHost { get; set; }
-    public string name { get; set; }
-
-    public Player(string name)
+    public void GenerateEquations(string gameId, string mode, int count)
     {
-        id = 0;
-        progress = 0;
-        score = 0;
-        isHost = false;
-        this.name = name;
-    }
-}
-
-public class GameMode
-{
-    public string type { get; set; }
-    public int count { get; set; }
-
-    public GameMode()
-    {
-        this.type = "time";
-        this.count = 100;
-    }
-
-    public GameMode(string type, int count)
-    {
-        this.type = type;
-        this.count = count;
-    }
-}
-
-public class Equation
-{
-    public int id { get; set; }
-    public string equation { get; set; }
-    public int answer { get; set; }
-
-    public Equation(int id, string equation, int answer)
-    {
-        this.id = id;
-        this.equation = equation;
-        this.answer = answer;
-    }
-}
-
-public class Game
-{
-    public string id { get; set; }
-    public GameMode gameMode { get; set; }
-    public Equation[] equations { get; set; }
-
-    public Game(string id, Equation[] eq, GameMode gameMode)
-    {
-        this.id = id;
-        this.gameMode = gameMode;
-        this.equations = eq;
+        Equation[] equations = Equation.GenerateAllEquations(count);
+        Game game = new Game(gameId, equations, new GameMode(mode, count));
+        games.Add(gameId, game);
     }
 }
